refactor: moved some UniverseView into MainWindow in ui.py
ui.py is a new file
